requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# AHS\loggers\core_logger.py: 4
PyYAML == 5.1.2

# AHS\main_spider.py: 4
jsonpath == 0.82

# AHS\Tools\tools.py: 9
lxml == 3.7.2

# AHS\Tools\excel_merge_tool.py: 3
# AHS\Tools\export_to_excel.py: 4
openpyxl == 2.6.3

# AHS\Tools\db_operator.py: 2
# AHS\Tools\export_to_excel.py: 5
# AHS\main_spider.py: 9
pymongo == 3.9.0

# AHS\Tools\tools.py: 11
# test.py: 2
requests == 2.20.0

